Upgrade Hipparchus to 3.1
EigenDecomposition() has been replaced by EigenDecompositionSymmetric(), however, this now introduces a MatrixUtils.checkSymmetric() call in the constructor, which breaks some of the current unittests since there are some double values very close to zero in a matrix that uses this decomposition.

The issue is due to the tolerance being in fact a relative tolerance check, not an absolute tolerance check. In this case the difference between the two entries 1.0e-16 and 1.0e-17 is 9.0e-17, and the relative check compares this to 1.0e-28 (which is max(1.0e-16, 1.0e-17) * 1.0e-12.

The workaround, as suggested by a Hipparchus dev, is to collapse small entries in the matrix before passing to EigenDecompositionSymmetric().

Also, a new ordering functionality (decreasing) is used in EigenDecompositionSymmetric()'s constructor which means SortedEigenDecompositionData() is not needed any more.

Refs:
	https://github.com/Hipparchus-Math/hipparchus/issues/365